Layout styles for new stage and guided new stage
