+++ src/components/Screens/NewStageScreen/NewStageScreen.js
@@ -6,23 +6,16 @@ import Screen from '@components/Screen/Screen';
 import interfaceTypes from './interfaceTypes';
 import CategorizedInterfaceList from './CategorizedInterfaceList';
 import InterfaceList from './InterfaceList';
-import Wizard from './Wizard';
 
 const animations = {
   show: { opacity: 1 },
   hide: { opacity: 0 },
 };
 
-const MODES = {
-  standard: 'STANDARD',
-  wizard: 'WIZARD',
-};
-
 const NewStageScreen = ({
   show,
   onComplete,
 }) => {
-  const [mode, setMode] = useState(MODES.standard);
   const [query, setQuery] = useState('');
 
   const buttons = useMemo(() => [
@@ -56,37 +49,30 @@ const NewStageScreen = ({
       type="new-stage"
     >
       <motion.div className="new-stage-screen">
-        { mode === MODES.wizard && <Wizard onQuit={() => setMode(MODES.standard)} /> }
-        { mode !== MODES.wizard && (
-          <>
-            <motion.div className="new-stage-screen__search">
-              <input
-                type="text"
-                value={query}
-                onChange={handleUpdateQuery}
-              />
-            </motion.div>
-            <motion.div className="new-stage-screen__container">
-              <CategorizedInterfaceList
-                onHelp={() => setMode(MODES.wizard)}
-              />
-              <AnimatePresence>
-                { query !== '' && (
-                  <motion.div
-                    className="new-stage-screen__results"
-                    variants={animations}
-                    initial="hide"
-                    exit="hide"
-                    animate="show"
-                    key="results"
-                  >
-                    <InterfaceList items={filteredInterfaces} />
-                  </motion.div>
-                )}
-              </AnimatePresence>
-            </motion.div>
-          </>
-        )}
+        <motion.div className="new-stage-screen__search">
+          <input
+            type="text"
+            value={query}
+            onChange={handleUpdateQuery}
+          />
+        </motion.div>
+        <motion.div className="new-stage-screen__container">
+          <CategorizedInterfaceList />
+          <AnimatePresence>
+            { query !== '' && (
+              <motion.div
+                className="new-stage-screen__results"
+                variants={animations}
+                initial="hide"
+                exit="hide"
+                animate="show"
+                key="results"
+              >
+                <InterfaceList items={filteredInterfaces} />
+              </motion.div>
+            )}
+          </AnimatePresence>
+        </motion.div>
       </motion.div>
     </Screen>
   );
