+++ src/styles/_components.scss
@@ -17,6 +17,7 @@
 @import './components/filter-group';
 @import './components/form';
 @import './components/grid';
+@import './components/guided-new-stage-screen';
 @import './components/home';
 @import './components/inline-edit-screen';
 @import './components/interview-script';
